requirements.txt
pandas
numpy
yfinance
tensorflow
scikit-learn
matplotlib
requests
python-dotenv
nltk
streamlit
plotly